+++ CHANGELOG.md
@@ -77,6 +77,10 @@ There are following change types:
 
 ## mini.clue
 
+### Evolve
+
+- Use current query clue (if not immediately after trigger) as a window title instead of showing keys verbatim. This shows extra context when navigating through mappings.
+
 ### Expand
 
 - Add `gen_clues.square_brackets` to generate clues for `[` and `]` keys. By @TheLeoP, PR #1937.